[#74][reservation] refactor: 패키지 이동 및 주석 추가
